+++ src/Query/Constraint/CriteriaTest.php
@@ -11,16 +11,20 @@
 
 declare(strict_types=1);
 
-namespace League\Csv\Constraint;
+namespace League\Csv\Query\Constraint;
+
 
 use ArrayIterator;
 use CallbackFilterIterator;
-use League\Csv\Extract;
+use League\Csv\Query\Select;
+use League\Csv\Query\QueryTestCase;
+
 use PHPUnit\Framework\Attributes\Test;
 
 use const ARRAY_FILTER_USE_BOTH;
 
-final class CriteriaTest extends ContraintTestCase
+final class CriteriaTest extends QueryTestCase
+
 {
     #[Test]
     public function it_returns_a_value_when_no_predicate_is_given(): void
@@ -34,8 +38,8 @@ final class CriteriaTest extends ContraintTestCase
     #[Test]
     public function it_returns_a_value_when_some_predicates_are_given(): void
     {
-        $predicate1 = fn (mixed $record, int $key) => Extract::value($record, 'volume') > 80;
-        $predicate2 = fn (mixed $record, int $key) => Extract::value($record, 'edition') < 6;
+        $predicate1 = fn (mixed $record, int $key) => Select::one($record, 'volume') > 80;
+        $predicate2 = fn (mixed $record, int $key) => Select::one($record, 'edition') < 6;
 
         self::assertSame([
             1 => ['volume' => 86, 'edition' => 1],
@@ -66,8 +70,8 @@ final class CriteriaTest extends ContraintTestCase
     {
         $data = new ArrayIterator($this->iterable);
 
-        $predicate1 = fn (mixed $record, int $key) => Extract::value($record, 'volume') > 80;
-        $predicate2 = fn (mixed $record, int $key) => Extract::value($record, 'edition') < 6;
+        $predicate1 = fn (mixed $record, int $key) => Select::one($record, 'volume') > 80;
+        $predicate2 = fn (mixed $record, int $key) => Select::one($record, 'edition') < 6;
 
         self::assertSame([
             0 => ['volume' => 67, 'edition' => 2],