+++ src/Query/Constraint/TwoColumnsTest.php
@@ -11,46 +11,38 @@
 
 declare(strict_types=1);
 
-namespace League\Csv\Constraint;
+namespace League\Csv\Query\Constraint;
 
 use League\Csv\StatementError;
 use PHPUnit\Framework\Attributes\Test;
+use League\Csv\Query\QueryTestCase;
 
-final class ColumnTest extends ContraintTestCase
+final class TwoColumnsTest extends QueryTestCase
+
 {
     #[Test]
     public function it_can_filter_the_tabular_data_based_on_the_column_value(): void
     {
-        $predicate = Column::filterOn('Country', '=', 'UK');
-        $result = $this->stmt->where($predicate)->process($this->document);
-
-        self::assertCount(1, $result);
-    }
-
-    #[Test]
-    public function it_can_filter_the_tabular_data_based_on_the_column_value_and_the_column_offset(): void
-    {
-        $predicate = Column::filterOn(0, 'in', ['1', '2']);
-        $result = $this->stmt->where($predicate)->process($this->document);
-
-        self::assertCount(2, $result);
-    }
-
-    #[Test]
-    public function it_can_not_filter_the_tabular_data_based_on_the_column_name(): void
-    {
-        $predicate = Column::filterOn('Country', '=', 'Country');
+        $predicate = TwoColumns::filterOn('Country', '=', 'City');
         $result = $this->stmt->where($predicate)->process($this->document);
 
         self::assertCount(0, $result);
     }
 
     #[Test]
+    public function it_can_filter_the_tabular_data_based_on_the_column_value_and_the_column_offset(): void
+    {
+        $predicate = TwoColumns::filterOn(0, 'contains', 0);
+        $result = $this->stmt->where($predicate)->process($this->document);
+
+        self::assertCount(5, $result);
+    }
+
     public function it_will_throw_if_the_column_does_not_exist(): void
     {
-        $predicate = Column::filterOn('Ville', '=', 'Dakar');
-        $this->expectExceptionObject(StatementError::dueToUnknownColumn('Ville', []));
+        $predicate = TwoColumns::filterOn('City', '=', 'Dakar');
+        $this->expectExceptionObject(StatementError::dueToUnknownColumn('City', []));
 
-        [...$this->stmt->where($predicate)->process($this->document)];
+        $this->stmt->where($predicate)->process($this->document);
     }
 }