+++ src/Query/PredicateCombinator.php
@@ -11,9 +11,11 @@
 
 declare(strict_types=1);
 
-namespace League\Csv\Constraint;
+namespace League\Csv\Query;
+
 
 use Closure;
+use League\Csv\Query\Predicate;
 
 /**
  * @phpstan-type Condition Predicate|Closure(mixed, array-key): bool