+++ src/Query/SortCombinator.php
@@ -11,9 +11,10 @@
 
 declare(strict_types=1);
 
-namespace League\Csv\Ordering;
+namespace League\Csv\Query;
 
 use Closure;
+use League\Csv\Query\Sort;
 
 /**
  * @phpstan-type Ordering Sort|Closure(mixed, mixed): int