+++ CHANGELOG.md
@@ -18,8 +18,7 @@ All Notable changes to `Csv` will be documented in this file
 - `Statement::whereColumnNot`
 - `Statement::orWhereColumn`
 - `Statement::xorWhereColumn`
-- `Constraint` feature to allow easier filtering of tabular data
-- `Ordering` feature to allow easier ordering of tabular data
+- `Query` feature to allow easier filtering, ordering and querying tabular data
 
 ### Deprecated